+++ packages/components/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### [Unreleased]
 
 ### Added
-- New `accessibility` property for `ARIA` attribute handling for: `Button`, `Button-Pure`, `Icon`, `Link`, `Link-Pure`, `Marque`, `Spinner`
+- New `aria` property for `ARIA` attribute handling for: `Button`, `Button-Pure`, `Icon`, `Link`, `Link-Pure`, `Marque`, `Spinner`
 
 #### Added
 - `unit` and `unitPosition` properties to `Text Field Wrapper`